fixed branch template and yml to azure
