+++ .github/workflows/main_neurplay-api.yml
@@ -23,7 +23,6 @@ jobs:
           
       - name: npm install, build
         run: |
-          cd backend
           npm install
 
       - name: Upload artifact for deployment job
