--- backend/nodemon.json
@@ -1,5 +1,0 @@
-{
-    "execMap": {
-        "js": "node -r sucrase/register"
-    }
-}
