+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+    "execMap": {
+        "js": "node -r sucrase/register"
+    }
+}
